fix(lens): the shared projection vocabulary
`@kolu/surface/verbs` stops being a topic bundle: what is not projection
vocabulary moves to its generative home, the volatile bridge gets a file of
its own, and the two rules both faces spell by hand become one derivation each.

- lowy F5 ∥ hickey F11 — `verbs.ts` is five concepts under a name that covers
  one: `messageOf` → `@kolu/surface/errors`, `SurfaceClientCallable` →
  `@kolu/surface/client`, neither re-exported (one concept, one import path)
- lowy F5 — the Schema→JSON-Schema bridge (an effect-version seam with a byte
  fixture) moves to `jsonSchemaBridge.ts`, re-exported from `verbs.ts` so no
  consumer import changes; its pins move with it
- hickey F12 — `messageOf` loses the third import path a day-old package minted
  (`./exit`'s and `index.ts`'s pass-through re-exports are gone)
- lowy F4 — "does this schema admit no argument?" was spelled twice across the
  two faces, held in agreement by a comment: one `admitsNoArgument`, both faces
  call in, each keeps its own policy for a `false`
- hickey F2 — the conservative `mutates` default was spelled character for
  character in both faces: one `exposureMutates` beside the `ToolExposure` it
  reads, because a safety default spelled twice can be relaxed in one place
- lowy F9 — "an owned connection to a served surface" was spelled a third time,
  with two incompatible `dispose` signatures: one `OwnedSurfaceConnection<C>`
  in `@kolu/surface/client`, aliased by both faces, so one host factory feeds
  both
- lowy F7 — the rules promoted to the shared home arrived without their pins:
  `decodeTextValue`'s two-reading contract, `messageOf`'s four fallbacks,
  `AdvertisedInput.inner` and `admitsNoArgument` now have theirs

Raised by the lowy ∥ hickey lens review. Not pushed or merged.
