+++ packages/surface/example/snippets/cli.ts
@@ -11,10 +11,13 @@
  */
 
 // #region imports
-import { buildSurfaceFace } from "@kolu/surface/client";
+import {
+  buildSurfaceFace,
+  type SurfaceClientCallable,
+} from "@kolu/surface/client";
 import { unixSocketLink } from "@kolu/surface/links/unix-socket";
 import { getRuntimeSocketPath } from "@kolu/surface/unix-socket";
-import type { SurfaceClientCallable, SurfaceVerb } from "@kolu/surface/verbs";
+import type { SurfaceVerb } from "@kolu/surface/verbs";
 import { type SurfaceCliConnection, surfaceCommands } from "@kolu/surface-cli";
 import { Command, Flag } from "effect/unstable/cli";
 // #endregion imports
